test: organizes test by class and adds one for mapping passage error
