+++ passageidentity/errors.py
@@ -4,6 +4,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+import typing_extensions
+
 if TYPE_CHECKING:
     from passageidentity.openapi_client.exceptions import ApiException
 
@@ -11,33 +13,34 @@ if TYPE_CHECKING:
 class PassageError(Exception):
     """Error class for handling Passage errors."""
 
+    @typing_extensions.deprecated(
+        "This should only be constructed by the Passage SDK. Use this type just for type checking.",
+    )
     def __init__(
         self,
         message: str,
         status_code: int | None = None,
         status_text: str | None = None,
         body: dict | None = None,
+        error_code: str | None = None,
     ) -> None:
         """Initialize the error with a message, status code, status text, and optional body."""
         self.message = message
         self.status_code = status_code
         self.status_text = status_text
+
+        self.error_code = error_code
+        self.error = None
+
         if body is not None:
             self.error = body["error"]
             self.error_code = body["code"]
-        else:
-            self.error = None
-            self.error_code = None
 
-    @staticmethod
-    def from_response_error(response_error: ApiException, message: str | None) -> PassageError:
+    @classmethod
+    def from_response_error(cls, response_error: ApiException, message: str | None = None) -> PassageError:
         """Initialize the error with a response body and optional message."""
         error_code = response_error.body["code"] if response_error.body else None
         error_msg = response_error.body["error"] if response_error.body else None
         msg = ": ".join(filter(None, [message, error_msg]))
 
-        psg_error = PassageError(msg)
-        psg_error.status_code = response_error.status
-        psg_error.error_code = error_code
-
-        return psg_error
+        return cls(message=msg, status_code=response_error.status, error_code=error_code)
